Making it not crash on Windows
