+++ containerd/main_windows.go
@@ -1,5 +1,21 @@
 package main
 
+import (
+	"os"
+
+	"github.com/codegangsta/cli"
+)
+
+var defaultStateDir = os.Getenv("PROGRAMDATA") + `\docker\containerd`
+
+const (
+	defaultListenType   = "tcp"
+	defaultGRPCEndpoint = "localhost:2377"
+)
+
+func appendPlatformFlags() {
+}
+
 // TODO Windows: May be able to factor out entirely
 func checkLimits() error {
 	return nil
@@ -7,4 +23,7 @@ func checkLimits() error {
 
 // No idea how to implement this on Windows.
 func reapProcesses() {
+}
+
+func setAppBefore(app *cli.App) {
 }
