+++ supervisor/machine_windows.go
@@ -1,0 +1,5 @@
+package supervisor
+
+func CollectMachineInformation() (Machine, error) {
+	return Machine{}, nil
+}
